requirements.txt
Flask
flask-cors
numpy
pandas
scikit-learn
lime
python-dotenv
psycopg2-binary
flask-sqlalchemy
